+++ src/components/Buttons/Redeem/RedeemWrappedToken.jsx
@@ -1,39 +1,52 @@
 import React, { useState, useEffect } from 'react';
-import { useSubstrate } from '../../substrate-lib';
-import { UNDERLYING_ASSETS_TYPES } from '../../util/constants';
+import { useSubstrate } from '../../../substrate-lib';
+import { WRAP_TOKEN_TYPES } from '../../../util/constants';
 
-import { Form, Dropdown, Button, Dimmer, Loader } from 'semantic-ui-react';
+import {
+	Form,
+	Input,
+	Dropdown,
+	Button,
+	Dimmer,
+	Loader,
+} from 'semantic-ui-react';
 
-function Redeem({ account }) {
+function RedeemWrappedToken({ account }) {
 	const { api, keyring } = useSubstrate();
+	const [amount, setAmount] = useState(0);
 	const [asset, setAsset] = useState('');
 	const [loading, setLoading] = useState(false);
 	const [isInvalid, setInvalid] = useState(true);
 
 	useEffect(() => {
-		setInvalid(!(asset && account));
-	}, [setInvalid, account, asset]);
+		setInvalid(!(asset && amount && account));
+	}, [setInvalid, account, amount, asset]);
 
 	const setInitialStates = () => {
+		setAmount(0);
 		setAsset('');
-		setInvalid(!(asset && account));
+		setInvalid(!(asset && amount && account));
 	};
 
-	const assets = UNDERLYING_ASSETS_TYPES.map((currency) => ({
+	const assets = WRAP_TOKEN_TYPES.map((currency) => ({
 		key: currency,
 		text: currency,
 		value: currency,
 	}));
 
+	const onChangeAmount = (e) => {
+		setAmount(e.target.value * 10 ** 18);
+	};
+
 	const onChangeAsset = (e) => {
 		setAsset(e.target.innerText);
 	};
 
-	const sendRedeemAll = async () => {
+	const redeemWrappedToken = async () => {
 		setLoading(true);
 		const currentUser = keyring.getPair(account);
 		await api.tx.minterestProtocol
-			.redeem(asset)
+			.redeemWrapped(asset, amount.toString())
 			.signAndSend(currentUser, ({ events = [], status }) => {
 				if (status.isFinalized) {
 					setLoading(false);
@@ -69,6 +82,11 @@ function Redeem({ account }) {
 
 	return (
 		<Form>
+			<Input
+				type='text'
+				placeholder='Enter the amount'
+				onChange={onChangeAmount}
+			/>
 			<Dropdown
 				placeholder='Asset'
 				search
@@ -78,14 +96,14 @@ function Redeem({ account }) {
 			/>
 			<Button
 				color={account ? 'green' : 'red'}
-				onClick={sendRedeemAll}
+				onClick={redeemWrappedToken}
 				disabled={isInvalid}
 			>
-				Redeem All Asset
+				Redeem Wrapped Token
 			</Button>
 			{isInvalid && <p>Please select to continue</p>}
 		</Form>
 	);
 }
 
-export default Redeem;
+export default RedeemWrappedToken;
